refactor(mongolian_ui): share FormDistrictCode between pos and stage ebarimt config, fix stage sheet reset
FormDistrictCode was duplicated identically between pos-in-ebarimt-config
and stage-in-ebarimt-config; moved the fixed version into settings/components
and re-export it from both FormFields barrels instead.

EditStageInEBarimtConfig had the same handleSubmit reset() bug as the pos
variant: calling reset() with no args after a successful save snapped the
sheet's visible fields back to pre-edit values while it was still closing.
