+++ frontend/plugins/mongolian_ui/src/modules/ebarimt/settings/components/FormFields/FormDistrictCode.tsx
