+++ frontend/plugins/mongolian_ui/src/modules/ebarimt/settings/stage-in-ebarimt-config/components/EditStageInEBarimtConfig.tsx
@@ -105,7 +105,6 @@ export const EditStageInEBarimtConfig = () => {
       );
       setOpen(null);
       setDetail(null);
-      reset();
     } catch {
       toast({
         title: t('error'),
