+++ frontend/plugins/mongolian_ui/src/modules/ebarimt/settings/stage-in-ebarimt-config/components/FormFields/index.ts
@@ -2,4 +2,4 @@ export { FormInput } from './FormInput';
 export { FormCheckbox } from './FormCheckbox';
 export { SelectEbarimtProductRules } from '../selects/SelectEbarimtProductRules';
 export { FormSelectEbarimtProductRules } from './FormSelectEbarimtProductRules';
-export { FormDistrictCode } from './FormDistrictCode';
+export { FormDistrictCode } from '@/ebarimt/settings/components/FormFields/FormDistrictCode';
